Remove unnecessary telemetry interfaces
After one year into telemetry, there is still no expected usage for
these.
