+++ telemetry/src/main/java/datadog/telemetry/TelemetryService.java
@@ -1,37 +1,257 @@
 package datadog.telemetry;
 
+import datadog.telemetry.api.AppDependenciesLoaded;
+import datadog.telemetry.api.AppIntegrationsChange;
+import datadog.telemetry.api.AppStarted;
 import datadog.telemetry.api.Dependency;
 import datadog.telemetry.api.DistributionSeries;
+import datadog.telemetry.api.Distributions;
+import datadog.telemetry.api.GenerateMetrics;
 import datadog.telemetry.api.Integration;
+import datadog.telemetry.api.KeyValue;
 import datadog.telemetry.api.LogMessage;
+import datadog.telemetry.api.Logs;
 import datadog.telemetry.api.Metric;
+import datadog.telemetry.api.Payload;
+import datadog.telemetry.api.RequestType;
+import datadog.trace.api.time.TimeSource;
+import java.util.Collections;
+import java.util.LinkedList;
+import java.util.List;
 import java.util.Map;
 import java.util.Queue;
+import java.util.concurrent.ArrayBlockingQueue;
+import java.util.concurrent.BlockingQueue;
+import java.util.concurrent.LinkedBlockingQueue;
+import java.util.function.Supplier;
 import okhttp3.Request;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
-public interface TelemetryService {
+public class TelemetryService {
 
-  // Special telemetry requests
-  void addStartedRequest();
+  private static final Logger log = LoggerFactory.getLogger(TelemetryService.class);
+  private static final String TELEMETRY_NAMESPACE_TAG_TRACER = "tracers";
 
-  Request appClosingRequest();
+  private static final int MAX_ELEMENTS_PER_REQUEST = 100;
 
-  // Data for periodic telemetry requests
-  boolean addConfiguration(Map<String, Object> configuration);
+  // https://github.com/DataDog/instrumentation-telemetry-api-docs/blob/main/GeneratedDocumentation/ApiDocs/v2/producing-telemetry.md#when-to-use-it-1
+  private static final int MAX_DEPENDENCIES_PER_REQUEST = 2000;
 
-  boolean addDependency(Dependency dependency);
+  private final Supplier<RequestBuilder> requestBuilderSupplier;
+  private final TimeSource timeSource;
+  private final int maxElementsPerReq;
+  private final int maxDepsPerReq;
+  private final int heartbeatIntervalMs;
+  private final int metricsIntervalMs;
+  private final BlockingQueue<KeyValue> configurations = new LinkedBlockingQueue<>();
+  private final BlockingQueue<Integration> integrations = new LinkedBlockingQueue<>();
+  private final BlockingQueue<Dependency> dependencies = new LinkedBlockingQueue<>();
+  private final BlockingQueue<Metric> metrics =
+      new LinkedBlockingQueue<>(1024); // recommended capacity?
 
-  boolean addIntegration(Integration integration);
+  private final BlockingQueue<LogMessage> logMessages = new LinkedBlockingQueue<>(1024);
 
-  boolean addMetric(Metric metric);
+  private final BlockingQueue<DistributionSeries> distributionSeries =
+      new LinkedBlockingQueue<>(1024);
 
-  boolean addLogMessage(LogMessage message);
+  private final Queue<Request> queue = new ArrayBlockingQueue<>(16);
 
-  boolean addDistributionSeries(DistributionSeries series);
+  private long lastHeartBeatTimestamp;
+  /*
+   * Keep track of Open Tracing and Open Telemetry integrations activation as they are mutually exclusive.
+   */
+  private boolean openTracingIntegrationEnabled;
+  private boolean openTelemetryIntegrationEnabled;
 
-  Queue<Request> prepareRequests();
+  public TelemetryService(
+      Supplier<RequestBuilder> requestBuilderSupplier,
+      TimeSource timeSource,
+      float heartBeatIntervalSec,
+      float metricsIntervalSec) {
+    this(
+        requestBuilderSupplier,
+        timeSource,
+        heartBeatIntervalSec,
+        metricsIntervalSec,
+        MAX_ELEMENTS_PER_REQUEST,
+        MAX_DEPENDENCIES_PER_REQUEST);
+  }
 
-  int getHeartbeatInterval();
+  // For testing purpose
+  TelemetryService(
+      Supplier<RequestBuilder> requestBuilderSupplier,
+      TimeSource timeSource,
+      float heartBeatIntervalSec,
+      float metricsIntervalSec,
+      int maxElementsPerReq,
+      int maxDepsPerReq) {
+    this.requestBuilderSupplier = requestBuilderSupplier;
+    this.timeSource = timeSource;
+    this.heartbeatIntervalMs = (int) (heartBeatIntervalSec * 1000); // we use time in milliseconds
+    this.metricsIntervalMs = (int) (metricsIntervalSec * 1000);
+    this.openTracingIntegrationEnabled = false;
+    this.openTelemetryIntegrationEnabled = false;
+    this.maxElementsPerReq = maxElementsPerReq;
+    this.maxDepsPerReq = maxDepsPerReq;
+  }
 
-  int getMetricsInterval();
+  public void addStartedRequest() {
+    Payload payload =
+        new AppStarted()
+            ._configuration(drainOrNull(configurations)) // absent if nothing
+            .integrations(drainOrEmpty(integrations, maxElementsPerReq)) // empty list if nothing
+            .dependencies(drainOrEmpty(dependencies, maxDepsPerReq)) // empty list if nothing
+            .requestType(RequestType.APP_STARTED);
+
+    queue.offer(requestBuilderSupplier.get().build(RequestType.APP_STARTED, payload));
+  }
+
+  public Request appClosingRequest() {
+    return requestBuilderSupplier.get().build(RequestType.APP_CLOSING);
+  }
+
+  public boolean addConfiguration(Map<String, Object> configuration) {
+    for (Map.Entry<String, Object> entry : configuration.entrySet()) {
+      if (!this.configurations.offer(new KeyValue().name(entry.getKey()).value(entry.getValue()))) {
+        return false;
+      }
+    }
+    return true;
+  }
+
+  public boolean addDependency(Dependency dependency) {
+    return this.dependencies.offer(dependency);
+  }
+
+  public boolean addIntegration(Integration integration) {
+    if ("opentelemetry-1".equals(integration.getName())) {
+      this.openTelemetryIntegrationEnabled = true;
+      warnAboutExclusiveIntegrations();
+    } else if ("opentracing".equals(integration.getName())) {
+      this.openTracingIntegrationEnabled = true;
+      warnAboutExclusiveIntegrations();
+    }
+    return this.integrations.offer(integration);
+  }
+
+  public boolean addMetric(Metric metric) {
+    return this.metrics.offer(metric);
+  }
+
+  public boolean addLogMessage(LogMessage message) {
+    return this.logMessages.offer(message);
+  }
+
+  public boolean addDistributionSeries(DistributionSeries series) {
+    return this.distributionSeries.offer(series);
+  }
+
+  public Queue<Request> prepareRequests() {
+    // New integrations
+    while (!integrations.isEmpty()) {
+      Payload payload =
+          new AppIntegrationsChange().integrations(drainOrEmpty(integrations, maxElementsPerReq));
+      Request request =
+          requestBuilderSupplier
+              .get()
+              .build(
+                  RequestType.APP_INTEGRATIONS_CHANGE,
+                  payload.requestType(RequestType.APP_INTEGRATIONS_CHANGE));
+      queue.offer(request);
+    }
+
+    // New dependencies
+    while (!dependencies.isEmpty()) {
+      Payload payload =
+          new AppDependenciesLoaded().dependencies(drainOrEmpty(dependencies, maxDepsPerReq));
+      Request request =
+          requestBuilderSupplier
+              .get()
+              .build(
+                  RequestType.APP_DEPENDENCIES_LOADED,
+                  payload.requestType(RequestType.APP_DEPENDENCIES_LOADED));
+      queue.offer(request);
+    }
+
+    // New metrics
+    while (!metrics.isEmpty()) {
+      Payload payload =
+          new GenerateMetrics()
+              .namespace(TELEMETRY_NAMESPACE_TAG_TRACER)
+              .series(drainOrEmpty(metrics, maxElementsPerReq));
+      Request request =
+          requestBuilderSupplier
+              .get()
+              .build(
+                  RequestType.GENERATE_METRICS, payload.requestType(RequestType.GENERATE_METRICS));
+      queue.offer(request);
+    }
+
+    // New messages
+    while (!logMessages.isEmpty()) {
+      Payload payload = new Logs().messages(drainOrEmpty(logMessages, maxElementsPerReq));
+      Request request =
+          requestBuilderSupplier
+              .get()
+              .build(RequestType.LOGS, payload.requestType(RequestType.LOGS));
+      queue.offer(request);
+    }
+
+    // New Distributions
+    while (!distributionSeries.isEmpty()) {
+      Payload payload =
+          new Distributions()
+              .namespace(TELEMETRY_NAMESPACE_TAG_TRACER)
+              .series(drainOrEmpty(distributionSeries, maxElementsPerReq));
+      Request request =
+          requestBuilderSupplier
+              .get()
+              .build(RequestType.DISTRIBUTIONS, payload.requestType(RequestType.DISTRIBUTIONS));
+      queue.offer(request);
+    }
+
+    // Heartbeat request if needed
+    final long curTime = this.timeSource.getCurrentTimeMillis();
+    if (curTime - lastHeartBeatTimestamp >= heartbeatIntervalMs) {
+      Request request = requestBuilderSupplier.get().build(RequestType.APP_HEARTBEAT);
+      queue.offer(request);
+      lastHeartBeatTimestamp = curTime;
+    }
+
+    return queue;
+  }
+
+  public int getHeartbeatInterval() {
+    return heartbeatIntervalMs;
+  }
+
+  public int getMetricsInterval() {
+    return metricsIntervalMs;
+  }
+
+  private void warnAboutExclusiveIntegrations() {
+    if (this.openTelemetryIntegrationEnabled && this.openTracingIntegrationEnabled) {
+      log.warn(
+          "Both Open Tracing and Open Telemetry integrations are enabled but mutually exclusive. Tracing performance can be degraded.");
+    }
+  }
+
+  private static <T> List<T> drainOrNull(BlockingQueue<T> srcQueue) {
+    return drainOrDefault(srcQueue, null, Integer.MAX_VALUE);
+  }
+
+  private static <T> List<T> drainOrEmpty(BlockingQueue<T> srcQueue, int maxItems) {
+    return drainOrDefault(srcQueue, Collections.<T>emptyList(), maxItems);
+  }
+
+  private static <T> List<T> drainOrDefault(
+      BlockingQueue<T> srcQueue, List<T> defaultList, int maxItems) {
+    List<T> list = new LinkedList<>();
+    int drained = srcQueue.drainTo(list, maxItems);
+    if (drained > 0) {
+      return list;
+    }
+    return defaultList;
+  }
 }
