+++ telemetry/src/main/java/datadog/telemetry/TelemetrySystem.java
@@ -4,7 +4,6 @@ import datadog.communication.ddagent.SharedCommunicationObjects;
 import datadog.telemetry.TelemetryRunnable.TelemetryPeriodicAction;
 import datadog.telemetry.dependency.DependencyPeriodicAction;
 import datadog.telemetry.dependency.DependencyService;
-import datadog.telemetry.dependency.DependencyServiceImpl;
 import datadog.telemetry.iast.IastTelemetryPeriodicAction;
 import datadog.telemetry.integration.IntegrationPeriodicAction;
 import datadog.telemetry.metric.WafMetricPeriodicAction;
@@ -29,7 +28,7 @@ public class TelemetrySystem {
 
   static DependencyService createDependencyService(Instrumentation instrumentation) {
     if (instrumentation != null && Config.get().isTelemetryDependencyServiceEnabled()) {
-      DependencyServiceImpl dependencyService = new DependencyServiceImpl();
+      DependencyService dependencyService = new DependencyService();
       dependencyService.installOn(instrumentation);
       dependencyService.schedulePeriodicResolution();
       return dependencyService;
@@ -63,7 +62,7 @@ public class TelemetrySystem {
       Instrumentation instrumentation, SharedCommunicationObjects sco) {
     DependencyService dependencyService = createDependencyService(instrumentation);
     TelemetryService telemetryService =
-        new TelemetryServiceImpl(
+        new TelemetryService(
             new RequestBuilderSupplier(sco.agentUrl),
             SystemTimeSource.INSTANCE,
             Config.get().getTelemetryHeartbeatInterval(),
