+++ telemetry/src/test/groovy/datadog/telemetry/dependency/LocationsCollectingTransformerSpecification.groovy
@@ -10,7 +10,7 @@ import java.util.concurrent.Executors
 
 class LocationsCollectingTransformerSpecification extends DepSpecification {
 
-  DependencyServiceImpl depService = new DependencyServiceImpl()
+  DependencyService depService = new DependencyService()
 
   LocationsCollectingTransformer transformer = new LocationsCollectingTransformer(depService)
 
@@ -93,7 +93,7 @@ class LocationsCollectingTransformerSpecification extends DepSpecification {
     }
 
     and:
-    def depService = Mock(DependencyServiceImpl)
+    def depService = Mock(DependencyService)
     def transformer = new LocationsCollectingTransformer(depService)
 
     when:
@@ -122,7 +122,7 @@ class LocationsCollectingTransformerSpecification extends DepSpecification {
     }
 
     and:
-    def depService = Mock(DependencyServiceImpl)
+    def depService = Mock(DependencyService)
     def transformer = new LocationsCollectingTransformer(depService)
 
     when:
@@ -159,7 +159,7 @@ class LocationsCollectingTransformerSpecification extends DepSpecification {
     ProtectionDomain domain = new ProtectionDomain(source, null)
 
     and:
-    def depService = Mock(DependencyServiceImpl)
+    def depService = Mock(DependencyService)
     def transformer = new LocationsCollectingTransformer(depService)
 
     when:
